requirements.txt
langchain
llama-index==0.6.15
openai
pandas
tiktoken
pypdf
pypdf4
python-dotenv
pymongo
weaviate-client
pinecone-client[grpc]
datasets
qwikidata